Change wind speed type to float and convert kelvin temp to celcius

+++ app/src/main/java/org/d3ifcool/hystorms/network/WeatherEntity.kt
@@ -24,6 +24,6 @@ data class MainDataWeather(
 )
 
 data class Wind(
-    val speed: Int,
+    val speed: Float,
     val deg: Int
 )
